requirements.txt
jsonlines~=4.0.0
tqdm~=4.66.1
pandas~=1.5.3
networkx==3.1
nltk~=3.8.1
Levenshtein==0.22.0
numpy==1.26.0
python-dateutil~=2.8.1
lightgbm~=3.3.5
thefuzz==0.20.0
streamlit==1.27.2
snapshottest~=0.6.0